Config rework
