--- config/encoder/encoder.go
@@ -1,8 +1,0 @@
-// Package encoder handles source encoding formats
-package encoder
-
-type Encoder interface {
-	Encode(interface{}) ([]byte, error)
-	Decode([]byte, interface{}) error
-	String() string
-}
