--- config/encoder/json/json.go
@@ -1,25 +1,0 @@
-package json
-
-import (
-	"encoding/json"
-
-	"github.com/micro/go-micro/v3/config/encoder"
-)
-
-type jsonEncoder struct{}
-
-func (j jsonEncoder) Encode(v interface{}) ([]byte, error) {
-	return json.Marshal(v)
-}
-
-func (j jsonEncoder) Decode(d []byte, v interface{}) error {
-	return json.Unmarshal(d, v)
-}
-
-func (j jsonEncoder) String() string {
-	return "json"
-}
-
-func NewEncoder() encoder.Encoder {
-	return jsonEncoder{}
-}
