--- config/source/file/format.go
@@ -1,15 +1,0 @@
-package file
-
-import (
-	"strings"
-
-	"github.com/micro/go-micro/v3/config/encoder"
-)
-
-func format(p string, e encoder.Encoder) string {
-	parts := strings.Split(p, ".")
-	if len(parts) > 1 {
-		return parts[len(parts)-1]
-	}
-	return e.String()
-}
