--- config/source/file/options.go
@@ -1,19 +1,0 @@
-package file
-
-import (
-	"context"
-
-	"github.com/micro/go-micro/v3/config/source"
-)
-
-type filePathKey struct{}
-
-// WithPath sets the path to file
-func WithPath(p string) source.Option {
-	return func(o *source.Options) {
-		if o.Context == nil {
-			o.Context = context.Background()
-		}
-		o.Context = context.WithValue(o.Context, filePathKey{}, p)
-	}
-}
